fix: 소비자 메인 페이지 상태 및 접근성 개선
