+++ src/components/consumer/PromotionCarousel.tsx
@@ -59,7 +59,7 @@ export function PromotionCarousel() {
       setCurrentIndex((prevIndex) =>
         prevIndex === promotionBanners.length - 1 ? 0 : prevIndex + 1
       );
-    }, 5000);
+    }, 10000);
 
     return () => window.clearInterval(timerId);
   }, []);
